ConstructStreamVAST This function constructs objects with class "StreamVAST". Depending on the arguments supplied, it will either initialize the object from countdata and reachdata, or it will add details to an already defined StreamVAST object.

ConstructStreamVAST This function constructs objects with class "StreamVAST". Depending on the arguments supplied, it will either initialize the object from countdata and reachdata, or it will add details to an already defined StreamVAST object.

Usage

ConstructStreamVAST(
  countdata,
  reachdata,
  surveydata,
  vastdata,
  vastmodel,
  countname,
  reachname,
  timetable,
  timescale,
  timerange,
  title,
  stats,
  preds,
  eval,
  spacedata,
  timedata,
  aucdata,
  unitconv,
  covariates
)

Arguments

countdata

A dataframe containing count data, such as outputed from AssembleReddData

reachdata

A sf object with data for reaches, such as from AssignReaches

surveydata

A optional data frame or sf object with survey information, such as from MakeSurveyTracks, doesn't do much right now

vastdata

A dataframe formatted specifically for use with VAST

vastmodel

A fitted VAST model

countname

A column name from countdata to use

reachname

a column name from reachdata to use

timetable

a dataframe converting timesteps to dates

timescale

The temporal resolution

timerange

A length two vector with the start and enddate for the temporal

title

A title or name for the stream system, use for plotting

stats

A list with summary stats, like RMSE,AIC

preds

A dataframe of all predictions from the VAST model

eval

A limited dataframe for evaluating model performance

spacedata

A dataframe with results aggregated by spatial frame

timedata

A dataframe with results aggregated by time frame

aucdata

a dataframe with Area under the Curve calculations derived from model

covariates

a list with elements for the covariates

StreamVAST

An StreamVAST object to append details

vastinput

A dataframe with spatial information for VAST, requires a specific format

vastnetwork

A dataframe with spatial information for VAST's stream functionality

vastnetworkLL

A dataframe with spatial information needed for VAST's stream functionality

vastsettings

a list of settings used to run a vast model

Value

A streamvast object with appropriate formatting