Point Setup and checking

Checks that a point is an sf object with the appropriate crs, or converts inputs into an sf object with POINT geometry; mostly used internally to check user inputs

Usage

PointSetup(point, crs)

Arguments

point

A sf, sfc, dataframe, or vector containing geometry or coordinates representing a point

crs

A crs to apply to the point, if it is not an sf or sfc object

Value

A sf object with POINT geometry