This function takes a streamvast object defines a temporal sampling frame and formats everything for use with vast

This function takes a streamvast object defines a temporal sampling frame and formats everything for use with vast

Usage

SetTemporalFrame(
  streamvast,
  startdate = NA,
  enddate = NA,
  Time = "Year",
  season,
  padyear = F,
  padhabitat = F
)

Arguments

streamvast

a streamvast object, minimally with countdata, reachdata defined

startdate

dates to truncate or expand the data in yyyy-mm-dd format

enddate

dates to truncate or expand the data in yyyy-mm-dd format

Time

character giving the timescale to use

season

a integer vector of length with order start Month, start day, end month, end day for a 'year' or 'season'

padyear

adds zeroes to the start and end of each runyear

padhabiat

adds zeros at regular intervals to areas designated non-habitat

Value

A streamvast object with formated with the appropriate temporal frame