Maps, the desired data, with options for facetting

Maps, the desired data, with options for facetting

Usage

plotPredictionMap(
  streamvast,
  mapvar = "pred_Density",
  facet = NA,
  FUN = "mean",
  background,
  subset,
  make.labels = F,
  xaxis.breaks = NA,
  yaxis.breaks = NA,
  palette = "turbo",
  legendname,
  max = Inf
)

Arguments

streamvast

a streamvast object with a prediction data frame

mapvar

the column in data to map

facet

a column to determine the facets, or a vector of values

FUN

a function like mean or median to summarize data

background

a sf object to put in the background

subset

a logical vector indicating which pred values to use

make.labels

should the reaches be labelled, turn off for multiple years

xaxis.breaks

Vector of x values for axis lines; useful to reduce crowding

yaxis.breaks

Vector of y values for axis lines; useful to reduce crowding

palette

a viridis color palette for plotting

legendname

character; a name for the legend

max

a maximum value to cap the scale; values above the max are set equal to the max

Value

A colored heatmap of the desired variable mapped over the stream network