Plot survival curves from a INLA survival model

Plot survival curves from a INLA survival model

Usage

plotSurvivalCurves(model, data, year = "all", reach = "all", title, mult = 1)

Arguments

model

a fitted INLA survival model

data

a dataframe of redd observations, usually from MakeReddSurvival

year

numeric; a subset of years to include in the plot

reach

numeric; a subset of reach numbers to include in the plot

title

character; a title to pass to ggtitle()

mult

a numeric multiplier, used to backtransform values, if necessary

Value

a ggplot showing the